requirements.txt
streamlit
pandas
streamlit-aggrid
python-dotenv